media, ie num fix, html validate, search fix

+++ js/shopping-bag.js
@@ -1,16 +1,27 @@
 "use sctrict";
 
+/* checking user agent for fixing elem.remove() in IE and EDGE */ 
+if (document.documentMode || /Edge/.test(navigator.userAgent)) {
+    var isIe = true;
+}
 
 var price = document.getElementById("basket-count");
 var content = document.getElementsByClassName("basket__table-body");
 var bagPrice = document.getElementsByClassName("basket__total-price");
 var form = document.getElementsByClassName("product__form");
-if (localStorage.getItem("basket") === null) {
-       localStorage.setItem("basket", []); 
+var orderBut = document.getElementsByClassName("basket__order-button");
+window.onload = function() {
+    if (localStorage.getItem("basket") === null) {
+        localStorage.setItem("basket", []); 
+    }
+    renderCount();
+    if (content[0] !== undefined){
+        renderBasket();
+    }
 }
-renderCount();
-renderBasket();
 
+
+/* get data from local storage */
 function getCartData(){
     try {
         JSON.parse(localStorage.getItem('basket'));
@@ -19,6 +30,8 @@ function getCartData(){
     }
     return JSON.parse(localStorage.getItem('basket'));
 }
+
+/*set data to local storage */
 
 function setCartData(o){
     localStorage.setItem('basket', JSON.stringify(o));
@@ -47,6 +60,13 @@ function toBag() {
     }
 
     if (document.body.contains(document.getElementById("error"))){
+        if (isIe) {
+            document.getElementById("error").remove = function() {
+                if (this.parentNode) {
+                    this.parentNode.removeChild(this);
+                }
+            };
+        }
         document.getElementById("error").remove();
     }
     
@@ -74,6 +94,8 @@ function toBag() {
     renderCount();
 }
 
+/* updating item counts */
+
 function renderCount() {
     let tmpArr = getCartData();
     let count = 0;
@@ -84,6 +106,8 @@ function renderCount() {
     return count;
 }
 
+/* to fill table rows */
+
 function renderBasket() {
     let tmpArr = getCartData();
     for (let i = 0; i < tmpArr.length; i++) {
@@ -91,6 +115,8 @@ function renderBasket() {
     }
     renderTotal();
 }
+
+/* render total sum */
 
 function renderTotal() {
     let total = 0;
@@ -102,7 +128,10 @@ function renderTotal() {
     bagPrice[0].innerText = total;
 }
 
+/* delete product from basket */
+
 function deleteProduct(id, size) {
+    rowId = -1;
     let tmpArr = getCartData();
     removed = [];
     for (let i = 0; i < tmpArr.length; i++) {
@@ -116,6 +145,8 @@ function deleteProduct(id, size) {
     renderCount();          
 }
 
+/* qty with count and total updating */
+
 function qtyChange(id, size, value) {
     let tmpArr = getCartData();
     for (let i = 0; i < tmpArr.length; i++) {
@@ -128,11 +159,77 @@ function qtyChange(id, size, value) {
     renderTotal();
 }
 
+/* qty for IE (use buttons instead of input number) */
+
+function qtyUp(id, size, rowId) {
+    let tmpArr = getCartData();
+    let input = document.getElementsByClassName("basket__input-number");
+    let itemTotal = document.getElementsByClassName("basket__price");
+    let total = parseInt(itemTotal[rowId].innerText.slice(1));
+    if (parseInt(input[rowId].value) >= 50) {
+        return false;
+    }
+    let tmp = parseInt(input[rowId].value) + 1;
+    let forOne;
+    for (let i = 0; i < catalog.length; i++) {
+        if (catalog[i].id === id) {
+            forOne = parseInt(catalog[i].price.split(" "));
+        }
+    }
+    itemTotal[rowId].innerText = "$" + (total + forOne);
+    input[rowId].value = tmp;
+    for (let i = 0; i < tmpArr.length; i++) {
+        if (tmpArr[i].id == id && tmpArr[i].size == size) {
+            tmpArr[i].count = parseInt(input[rowId].value);
+        }
+    }
+    setCartData(tmpArr);
+    renderCount();
+    renderTotal();
+}
+
+function qtyDown(id, size, rowId) {
+    let tmpArr = getCartData();
+    let input = document.getElementsByClassName("basket__input-number");
+    let itemTotal = document.getElementsByClassName("basket__price");
+    let total = parseInt(itemTotal[rowId].innerText.slice(1));
+    if (parseInt(input[rowId].value) <= 1) {
+        return false;
+    }
+    let tmp = parseInt(input[rowId].value) - 1;
+  
+    let forOne;
+    for (let i = 0; i < catalog.length; i++) {
+        if (catalog[i].id === id) {
+            forOne = parseInt(catalog[i].price.split(" "));
+        }
+    }
+    itemTotal[rowId].innerText = "$" + (total - forOne);
+    input[rowId].value = tmp;
+    for (let i = 0; i < tmpArr.length; i++) {
+        if (tmpArr[i].id == id && tmpArr[i].size == size) {
+            tmpArr[i].count = parseInt(input[rowId].value);
+        }
+    }
+    setCartData(tmpArr);
+    renderCount();
+    renderTotal();
+}
+
+/* make order button */
+
 function makeOrder() {
+    if (localStorage.getItem("basket") === null || localStorage.getItem("basket") === [] || localStorage.getItem("basket") === "") {
+        orderBut[0].innerText = "BASKET IS EMPTY!"
+        orderBut[0].disabled = true;
+        return false;
+    }
     localStorage.clear();
     window.location.href = 'thank-you.html';
 }
 
+/* basket row creating */
+var rowId = -1;
 
 function basketRow(id, chosedSize, count) {
     let price;
@@ -145,6 +242,8 @@ function basketRow(id, chosedSize, count) {
             img = catalog[i].image;
         }
     }
+    rowId++;
+    console.log(rowId);
     return [
       '<tr class="basket__table-row" data-id=' + id + ' data-size=' + chosedSize + '>',
         '<td class="basket__table-cell">',
@@ -161,17 +260,19 @@ function basketRow(id, chosedSize, count) {
         '<td class="basket__table-cell">' + "One color" + '</td>',
         '<td class="basket__table-cell">' + chosedSize + '</td>',
         '<td class="basket__table-cell">',
-          '<input type="number" class="basket__input-number" value="' + count + '" min="1" max="30" onkeypress="return false" onchange="qtyChange('+ 
-          id + ',' + chosedSize + ',' + 'this.value)">',
+          '<input type="number" class="basket__input-number" value="' + count + '" min="1" max="30"  onkeypress="return false" onchange="qtyChange('+ 
+          id + ',' + chosedSize + ',' + 'this.value)" readonly>',
+          '<button id="arrowUp" style="font-size: 20px; cursor: pointer; margin-left: -10px; margin-top: -5px;" onclick="qtyDown('+ id + ',' + chosedSize + ',' + rowId + ')">&#x25c2</button>', 
+          '<button id="arrowDown" style="font-size: 20px; cursor: pointer; margin-left: 5px; margin-top: -5px;" onclick="qtyUp('+ id + ',' + chosedSize + ',' + rowId + ')">&#x25b8</button>',
         '</td>',
         '<td class="basket__table-cell">',
           '<span class="basket__price">' + "$" + price +  '</span>',
         '</td>',
         '<td class="basket__table-cell">',
-          '<button class="basket__delete-button" onclick="deleteProduct(' + id + ',' + chosedSize + ')">',
+          '<button class="basket__delete-button" style="cursor:pointer;" onclick="deleteProduct(' + id + ',' + chosedSize + ')">',
             '&#10060;',
           '</button>',
         '</td>',
       '</tr>',
     ].join('');
-  }
+}
